pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ldf"
version = "0.1.0"
description = "LLM Development Framework - Spec-driven development for AI-assisted engineering"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "LLM.info Team"},
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["development", "framework", "spec-driven", "ai", "mcp"]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
mcp = [
    "mcp>=0.9.0",
]
automation = [
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
]

[project.scripts]
ldf = "ldf.cli:main"

[project.urls]
Homepage = "https://github.com/LLMdotInfo/ldf"
Documentation = "https://github.com/LLMdotInfo/ldf#readme"
Repository = "https://github.com/LLMdotInfo/ldf"
Issues = "https://github.com/LLMdotInfo/ldf/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ldf*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=ldf --cov-report=term-missing"
